unsigned int <-> int の比較を行わないように

+++ src/ofxIcon.cpp
@@ -5,17 +5,17 @@
 namespace ofxIcon {
 namespace util {
 void fillColor(ofPixels & pixels, ofColor color) {
-	for (int x = 0; x < pixels.getWidth(); x++) {
-		for (int y = 0; y < pixels.getHeight(); y++) {
+	for (int x = 0; x < static_cast<int>(pixels.getWidth()); x++) {
+		for (int y = 0; y < static_cast<int>(pixels.getHeight()); y++) {
 			pixels.setColor(x, y, color);
 		}
 	}
 }
 
 void fillGrad(ofPixels & pixels, ofColor fillColor0, ofColor fillColor1, ofColor fillColor2) {
-	int baseline = static_cast<int>(static_cast<float>(pixels.getHeight()) * 0.4f);
-	for (int x = 0; x < pixels.getWidth(); x++) {
-		for (int y = 0; y < pixels.getHeight(); y++) {
+	int baseline = static_cast<int>(static_cast<float>(static_cast<int>(pixels.getHeight())) * 0.4f);
+	for (int x = 0; x < static_cast<int>(pixels.getWidth()); x++) {
+		for (int y = 0; y < static_cast<int>(pixels.getHeight()); y++) {
 			ofColor fc0 = fillColor0;
 			ofColor fc1 = fillColor1;
 			ofColor fc2 = fillColor2;
@@ -25,7 +25,7 @@ void fillGrad(ofPixels & pixels, ofColor fillColor0, ofColor fillColor1, ofColor
 			}
 			else if (y >= baseline) {
 				int start = y - baseline;
-				int end = pixels.getHeight() - baseline;
+				int end = static_cast<int>(pixels.getHeight()) - baseline;
 				float f = static_cast<float>(start) / static_cast<float>(end);
 				pixels.setColor(x, y, fc1.lerp(fc2, f));
 			}
@@ -34,23 +34,23 @@ void fillGrad(ofPixels & pixels, ofColor fillColor0, ofColor fillColor1, ofColor
 }
 
 void drawBorder(ofPixels & pixels, ofColor color, int edgeSize) {
-	for (int x = 0; x < pixels.getWidth(); x++) {
+	for (int x = 0; x < static_cast<int>(pixels.getWidth()); x++) {
 		for (int y = 0; y < edgeSize; y++) {
 			pixels.setColor(x, y, color);
 		}
 	}
-	for (int y = 0; y < pixels.getHeight(); y++) {
-		for (int x = pixels.getWidth() - edgeSize; x < pixels.getWidth(); x++) {
+	for (int y = 0; y < static_cast<int>(pixels.getHeight()); y++) {
+		for (int x = static_cast<int>(pixels.getWidth()) - edgeSize; x < static_cast<int>(pixels.getWidth()); x++) {
 			pixels.setColor(x, y, color);
 		}
 	}
 	for (int x = 0; x < edgeSize; x++) {
-		for (int y = 0; y < pixels.getHeight(); y++) {
+		for (int y = 0; y < static_cast<int>(pixels.getHeight()); y++) {
 			pixels.setColor(x, y, color);
 		}
 	}
-	for (int x = 0; x < pixels.getWidth(); x++) {
-		for (int y = pixels.getHeight() - edgeSize; y < pixels.getHeight(); y++) {
+	for (int x = 0; x < static_cast<int>(pixels.getWidth()); x++) {
+		for (int y = static_cast<int>(pixels.getHeight()) - edgeSize; y < static_cast<int>(pixels.getHeight()); y++) {
 			pixels.setColor(x, y, color);
 		}
 	}
@@ -175,23 +175,23 @@ InputFieldStyle InputFieldStyle::borderLightColor(ofColor _borderLightColor) {
 }
 void writeInputFieldImage(ofPixels & pixels, InputFieldStyle style) {
 	util::fillColor(pixels, style._fillColor);
-	for (int x = 0; x < pixels.getWidth(); x++) {
+	for (int x = 0; x < static_cast<int>(pixels.getWidth()); x++) {
 		for (int y = 0; y < style._edgeSize; y++) {
 			pixels.setColor(x, y, style._borderShadowColor);
 		}
 		for (int y = style._edgeSize; y < style._edgeSize*2; y++) {
 			pixels.setColor(x, y, style._borderLightColor);
 		}
-		for (int y = pixels.getHeight() - (style._edgeSize * 2); y < pixels.getHeight() - style._edgeSize; y++) {
+		for (int y = static_cast<int>(pixels.getHeight()) - (style._edgeSize * 2); y < static_cast<int>(pixels.getHeight()) - style._edgeSize; y++) {
 			pixels.setColor(x, y, style._borderShadowColor);
 		}
-		for (int y = pixels.getHeight() - style._edgeSize; y < pixels.getHeight(); y++) {
+		for (int y = static_cast<int>(pixels.getHeight()) - style._edgeSize; y < static_cast<int>(pixels.getHeight()); y++) {
 			pixels.setColor(x, y, style._borderLightColor);
 		}
 	}
-	for (int y = 0; y < pixels.getHeight(); y++) {
+	for (int y = 0; y < static_cast<int>(pixels.getHeight()); y++) {
 		for (int x = 0; x < style._edgeSize; x++) {
-			if (y >= pixels.getHeight() - style._edgeSize * 2) {
+			if (y >= static_cast<int>(pixels.getHeight()) - style._edgeSize * 2) {
 				break;
 			}
 			pixels.setColor(x, y, style._borderShadowColor);
@@ -200,18 +200,18 @@ void writeInputFieldImage(ofPixels & pixels, InputFieldStyle style) {
 			if (y < style._edgeSize) {
 				break;
 			}
-			if (y >= pixels.getHeight() - style._edgeSize * 2) {
+			if (y >= static_cast<int>(pixels.getHeight()) - style._edgeSize * 2) {
 				break;
 			}
 			pixels.setColor(x, y, style._borderLightColor);
 		}
-		for (int x = pixels.getWidth() - (style._edgeSize * 2); x < pixels.getWidth() - style._edgeSize; x++) {
-			if (y >= pixels.getHeight() - style._edgeSize) {
+		for (int x = static_cast<int>(pixels.getWidth()) - (style._edgeSize * 2); x < static_cast<int>(pixels.getWidth()) - style._edgeSize; x++) {
+			if (y >= static_cast<int>(pixels.getHeight()) - style._edgeSize) {
 				break;
 			}
 			pixels.setColor(x, y, style._borderShadowColor);
 		}
-		for (int x = pixels.getWidth() - style._edgeSize; x < pixels.getWidth(); x++) {
+		for (int x = static_cast<int>(pixels.getWidth()) - style._edgeSize; x < static_cast<int>(pixels.getWidth()); x++) {
 			pixels.setColor(x, y, style._borderLightColor);
 		}
 	}
@@ -288,13 +288,13 @@ int DropdownStyle::computeSlotWidth(int width) const {
 
 void writeDropdownImage(ofPixels & pixels, DropdownStyle style) {
 	util::fillColor(pixels, style._fillColor);
-	int baseline = static_cast<int>(static_cast<float>(pixels.getHeight()) * 0.4f);
-	int slotSizeX = style.computeSlotWidth(pixels.getWidth());
-	int slotSizeY = pixels.getHeight();
-	int slotPosX = pixels.getWidth() - slotSizeX;
+	int baseline = static_cast<int>(static_cast<float>(static_cast<int>(pixels.getHeight())) * 0.4f);
+	int slotSizeX = style.computeSlotWidth(static_cast<int>(pixels.getWidth()));
+	int slotSizeY = static_cast<int>(pixels.getHeight());
+	int slotPosX = static_cast<int>(pixels.getWidth()) - slotSizeX;
 	int slotPosY = 0;
-	for (int x = slotPosX; x < pixels.getWidth(); x++) {
-		for (int y = 0; y < pixels.getHeight(); y++) {
+	for (int x = slotPosX; x < static_cast<int>(pixels.getWidth()); x++) {
+		for (int y = 0; y < static_cast<int>(pixels.getHeight()); y++) {
 			ofColor fillColor0 = style._fillSlotColor0;
 			ofColor fillColor1 = style._fillSlotColor1;
 			ofColor fillColor2 = style._fillSlotColor2;
@@ -304,7 +304,7 @@ void writeDropdownImage(ofPixels & pixels, DropdownStyle style) {
 			}
 			else if (y >= baseline) {
 				int start = y - baseline;
-				int end = pixels.getHeight() - baseline;
+				int end = static_cast<int>(pixels.getHeight()) - baseline;
 				float f = static_cast<float>(start) / static_cast<float>(end);
 				pixels.setColor(x, y, fillColor1.lerp(fillColor2, f));
 			}
@@ -326,7 +326,7 @@ void writeDropdownImage(ofPixels & pixels, DropdownStyle style) {
 	}
 	util::drawBorder(pixels, style._borderColor, style._edgeSize);
 	for (int x = slotPosX; x < slotPosX + style._edgeSize; x++) {
-		for (int y = 0; y < pixels.getHeight(); y++) {
+		for (int y = 0; y < static_cast<int>(pixels.getHeight()); y++) {
 			pixels.setColor(x, y, style._borderColor);
 		}
 	}
@@ -382,13 +382,13 @@ void writeCheckBoxImage(ofPixels & pixels, CheckBoxStyle style) {
 		.fillColor2(style._fillColor2));
 	if (style._selected) {
 		glm::ivec2 p0(
-			(pixels.getWidth() / 2) - pixels.getWidth() / 4,pixels.getHeight() / 2
+			(static_cast<int>(pixels.getWidth()) / 2) - static_cast<int>(pixels.getWidth()) / 4,static_cast<int>(pixels.getHeight()) / 2
 		);
 		glm::ivec2 p1(
-			(pixels.getWidth() / 2) - pixels.getWidth() / 4, (pixels.getHeight() / 2) + pixels.getHeight() / 4
+			(static_cast<int>(pixels.getWidth()) / 2) - static_cast<int>(pixels.getWidth()) / 4, (static_cast<int>(pixels.getHeight()) / 2) + static_cast<int>(pixels.getHeight()) / 4
 		);
 		glm::ivec2 p2(
-			(pixels.getWidth() / 2) + pixels.getWidth() / 4, (pixels.getHeight() / 2) - pixels.getHeight() / 4
+			(static_cast<int>(pixels.getWidth()) / 2) + static_cast<int>(pixels.getWidth()) / 4, (static_cast<int>(pixels.getHeight()) / 2) - static_cast<int>(pixels.getHeight()) / 4
 		);
 		util::drawLine(pixels, style._lineColor, p0, p1);
 		util::drawLine(pixels, style._lineColor, p0 + glm::ivec2(1,0), p1 + glm::ivec2(1, 0));
@@ -444,11 +444,11 @@ RadioButtonStyle RadioButtonStyle::fillColor2(ofColor _fillColor2) {
 	return c;
 }
 void writeRadioButtonImage(ofPixels & pixels, RadioButtonStyle style) {
-	int radiusX = pixels.getWidth() / 2;
-	int radiusY = pixels.getHeight() / 2;
+	int radiusX = static_cast<int>(pixels.getWidth()) / 2;
+	int radiusY = static_cast<int>(pixels.getHeight()) / 2;
 	int centerX = radiusX;
 	int centerY = radiusY;
-	int baseline = static_cast<int>(static_cast<float>(pixels.getHeight()) * 0.4f);
+	int baseline = static_cast<int>(static_cast<float>(static_cast<int>(pixels.getHeight())) * 0.4f);
 	util::fillColor(pixels, ofColor(0, 0, 0, 0));
 	for (int i = 0; i < 360; i++) {
 		float f = static_cast<float>(i);
@@ -465,7 +465,7 @@ void writeRadioButtonImage(ofPixels & pixels, RadioButtonStyle style) {
 		int innerix = static_cast<int>(innerfx);
 		int inneriy = static_cast<int>(innerfy);
 		int startX = std::max(0, centerX - innerix);
-		int endX = std::min(centerX + innerix, static_cast<int>(pixels.getWidth())-1);
+		int endX = std::min(centerX + innerix, static_cast<int>(static_cast<int>(pixels.getWidth()))-1);
 		for (int x = startX; x < endX; x++) {
 			pixels.setColor(x, inneriy, ofColor::white);
 			int y = inneriy;
@@ -477,7 +477,7 @@ void writeRadioButtonImage(ofPixels & pixels, RadioButtonStyle style) {
 				pixels.setColor(x, y, fillColor0.lerp(fillColor1, f));
 			} else if (y >= baseline) {
 				int start = y - baseline;
-				int end = pixels.getHeight() - baseline;
+				int end = static_cast<int>(pixels.getHeight()) - baseline;
 				float f = static_cast<float>(start) / static_cast<float>(end);
 				pixels.setColor(x, y, fillColor1.lerp(fillColor2, f));
 			}
@@ -486,7 +486,7 @@ void writeRadioButtonImage(ofPixels & pixels, RadioButtonStyle style) {
 	if (!style._selected) {
 		return;
 	}
-	int circleSize = ((pixels.getWidth() + pixels.getHeight()) / 2) / 3;
+	int circleSize = ((static_cast<int>(pixels.getWidth()) + static_cast<int>(pixels.getHeight())) / 2) / 3;
 	for (int i = 0; i < 360; i++) {
 		float f = static_cast<float>(i);
 		f = f * (3.14f / 180.0f);
@@ -575,13 +575,13 @@ SliderKnobStyle SliderKnobStyle::fillColor2(ofColor _fillColor2) {
 }
 void writeSliderKnobImage(ofPixels & pixels, SliderKnobStyle style) {
 	util::fillColor(pixels, ofColor(0, 0, 0, 0));
-	int hw = pixels.getWidth() / 2;
-	int hh = pixels.getHeight() / 2;
+	int hw = static_cast<int>(pixels.getWidth()) / 2;
+	int hh = static_cast<int>(pixels.getHeight()) / 2;
 	for (int i = 0; i < style._edgeSize; i++) {
 		util::drawLine(pixels,
 			style._borderColor,
 			glm::ivec2(0, i),
-			glm::ivec2(pixels.getWidth(), i)
+			glm::ivec2(static_cast<int>(pixels.getWidth()), i)
 		);
 		util::drawLine(pixels,
 			style._borderColor,
@@ -590,34 +590,34 @@ void writeSliderKnobImage(ofPixels & pixels, SliderKnobStyle style) {
 		);
 		util::drawLine(pixels,
 			style._borderColor,
-			glm::ivec2(pixels.getWidth()-(i+1), 0),
-			glm::ivec2(pixels.getWidth()-(i+1), hh)
+			glm::ivec2(static_cast<int>(pixels.getWidth())-(i+1), 0),
+			glm::ivec2(static_cast<int>(pixels.getWidth())-(i+1), hh)
 		);
 		util::drawLine(pixels,
 			style._borderColor,
 			glm::ivec2(i, hh),
-			glm::ivec2(hw- (style._edgeSize - i), pixels.getHeight())
+			glm::ivec2(hw- (style._edgeSize - i), static_cast<int>(pixels.getHeight()))
 		);
 		util::drawLine(pixels,
 			style._borderColor,
-			glm::ivec2(pixels.getWidth() - (i + 1), hh),
-			glm::ivec2(hw+(style._edgeSize-i), pixels.getHeight())
+			glm::ivec2(static_cast<int>(pixels.getWidth()) - (i + 1), hh),
+			glm::ivec2(hw+(style._edgeSize-i), static_cast<int>(pixels.getHeight()))
 		);
 		util::drawLine(pixels,
 			style._borderColor,
-			glm::ivec2(hw - (style._edgeSize), pixels.getHeight() - (i+1)),
-			glm::ivec2(hw + (style._edgeSize), pixels.getHeight() - (i+1))
+			glm::ivec2(hw - (style._edgeSize), static_cast<int>(pixels.getHeight()) - (i+1)),
+			glm::ivec2(hw + (style._edgeSize), static_cast<int>(pixels.getHeight()) - (i+1))
 		);
 	}
-	int baseline = static_cast<int>(static_cast<float>(pixels.getHeight()) * 0.4f);
-	for (int y = 0; y < pixels.getHeight(); y++) {
+	int baseline = static_cast<int>(static_cast<float>(static_cast<int>(pixels.getHeight())) * 0.4f);
+	for (int y = 0; y < static_cast<int>(pixels.getHeight()); y++) {
 		if (y < style._edgeSize) {
 			continue;
 		}
 		int startX = 0; bool findStartX = false;
 		int endX = 0;   bool findEndX = false;
 		// ŠJŽn, I—¹‚ðŽæ“¾‚·‚é
-		for (int x = 0; x < pixels.getWidth(); x++) {
+		for (int x = 0; x < static_cast<int>(pixels.getWidth()); x++) {
 			ofColor c = pixels.getColor(x, y);
 			if (!findStartX && c == style._borderColor) {
 				findStartX = true;
@@ -627,7 +627,7 @@ void writeSliderKnobImage(ofPixels & pixels, SliderKnobStyle style) {
 				break;
 			}
 		}
-		for (int x = pixels.getWidth()-1; x >=0; x--) {
+		for (int x = static_cast<int>(pixels.getWidth())-1; x >=0; x--) {
 			ofColor c = pixels.getColor(x, y);
 			if (!findEndX && c == style._borderColor) {
 				findEndX = true;
@@ -647,7 +647,7 @@ void writeSliderKnobImage(ofPixels & pixels, SliderKnobStyle style) {
 			}
 			else if (y >= baseline) {
 				int start = y - baseline;
-				int end = pixels.getHeight() - baseline;
+				int end = static_cast<int>(pixels.getHeight()) - baseline;
 				float f = static_cast<float>(start) / static_cast<float>(end);
 				pixels.setColor(x, y, fc1.lerp(fc2, f));
 			}
